% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_citation.R
\name{convert_reference}
\alias{convert_reference}
\title{content <- c(
  "blah (Wulder et al. 2012; White et al. 2013; Beland et al. 2019) blah",
  "blah blah (Ellsworth and Reich 1993; White et al. 2013) blah",
  "blah blah (Beland et al. 2019; Crespo-Peremarch et al. 2020) blah",
  "blah blah (White et al. 2013) blah",
  "blah blah (Ellsworth and Reich 1993) blah",
  "(MacArthur and Horn 1969; Aber 1979; Bassow and\nBazzaz 1997; Tackenberg 2007)",
  "nope (nope not this one) here"
  )
expect_equal(convert_selection(content), gsub("\n", " ", c("blah [@wulder_etal12;}
\usage{
convert_reference(x)
}
\description{
content <- c(
  "blah (Wulder et al. 2012; White et al. 2013; Beland et al. 2019) blah",
  "blah blah (Ellsworth and Reich 1993; White et al. 2013) blah",
  "blah blah (Beland et al. 2019; Crespo-Peremarch et al. 2020) blah",
  "blah blah (White et al. 2013) blah",
  "blah blah (Ellsworth and Reich 1993) blah",
  "(MacArthur and Horn 1969; Aber 1979; Bassow and\nBazzaz 1997; Tackenberg 2007)",
  "nope (nope not this one) here"
  )
expect_equal(convert_selection(content), gsub("\n", " ", c("blah [@wulder_etal12;
}
