% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_n_if_needed}
\alias{sample_n_if_needed}
\title{fill rows with sampled values (upsampling)
If the data frame has more than n rows, then sample through the data frame without replacement
If}
\usage{
sample_n_if_needed(tbl, size, fill)
}
\description{
fill rows with sampled values (upsampling)
If the data frame has more than n rows, then sample through the data frame without replacement
If
}
